Renamed ClientModel to ClientModels as to not conflict with the ClientModel class inside that namespace.
