+++ src/AutoRest.CSharp.V3/ClientModels/BinaryTypeReference.cs
@@ -1,17 +1,12 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-using AutoRest.CSharp.V3.Pipeline.Generated;
-
-namespace AutoRest.CSharp.V3.ClientModel
+namespace AutoRest.CSharp.V3.ClientModels
 {
-    internal class SchemaTypeReference: ClientTypeReference
+    internal class BinaryTypeReference : ClientTypeReference
     {
-        public Schema Schema { get; }
-
-        public SchemaTypeReference(Schema schema, bool isNullable)
+        public BinaryTypeReference(bool isNullable)
         {
-            Schema = schema;
             IsNullable = isNullable;
         }
