+++ src/AutoRest.CSharp.V3/ClientModels/ClientConstant.cs
@@ -3,7 +3,7 @@
 
 using System.Diagnostics;
 
-namespace AutoRest.CSharp.V3.ClientModel
+namespace AutoRest.CSharp.V3.ClientModels
 {
     internal struct ClientConstant
     {
