+++ src/AutoRest.CSharp.V3/ClientModels/ClientModel.cs
@@ -1,10 +1,10 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-namespace AutoRest.CSharp.V3.ClientModel
+namespace AutoRest.CSharp.V3.ClientModels
 {
-    internal abstract class ClientTypeReference
+    internal abstract class ClientModel
     {
-        public abstract bool IsNullable { get; }
+        public abstract string Name { get; }
     }
 }
