+++ src/AutoRest.CSharp.V3/ClientModels/ClientObject.cs
@@ -4,7 +4,7 @@
 using System.Collections.Generic;
 using AutoRest.CSharp.V3.Pipeline.Generated;
 
-namespace AutoRest.CSharp.V3.ClientModel
+namespace AutoRest.CSharp.V3.ClientModels
 {
     internal class ClientObject : ClientModel, ISchemaTypeProvider
     {
