+++ src/AutoRest.CSharp.V3/ClientModels/ClientObjectConstant.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-namespace AutoRest.CSharp.V3.ClientModel
+namespace AutoRest.CSharp.V3.ClientModels
 {
     internal class ClientObjectConstant
     {
