+++ src/AutoRest.CSharp.V3/ClientModels/ClientTypeReference.cs
@@ -1,10 +1,10 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-namespace AutoRest.CSharp.V3.ClientModel
+namespace AutoRest.CSharp.V3.ClientModels
 {
-    internal abstract class ClientModel
+    internal abstract class ClientTypeReference
     {
-        public abstract string Name { get; }
+        public abstract bool IsNullable { get; }
     }
 }
