+++ src/AutoRest.CSharp.V3/ClientModels/ConstantOrParameter.cs
@@ -3,7 +3,7 @@
 
 using System;
 
-namespace AutoRest.CSharp.V3.ClientModel
+namespace AutoRest.CSharp.V3.ClientModels
 {
     internal struct ConstantOrParameter
     {
