+++ src/AutoRest.CSharp.V3/ClientModels/ServiceClient.cs
@@ -1,17 +1,18 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-namespace AutoRest.CSharp.V3.ClientModel
+namespace AutoRest.CSharp.V3.ClientModels
 {
-    internal class ClientEnumValue
+    internal class ServiceClient
     {
-        public ClientEnumValue(string name, ClientConstant value)
+        public ServiceClient(string name, ClientMethod[] methods)
         {
             Name = name;
-            Value = value;
+            Methods = methods;
         }
 
         public string Name { get; }
-        public ClientConstant Value { get; }
+
+        public ClientMethod[] Methods { get; }
     }
 }
