+++ src/AutoRest.CSharp.V3/Pipeline/Extensions.cs
@@ -3,7 +3,7 @@
 
 using System;
 using System.Collections.Generic;
-using AutoRest.CSharp.V3.ClientModel;
+using AutoRest.CSharp.V3.ClientModels;
 using AutoRest.CSharp.V3.CodeGen;
 using AutoRest.CSharp.V3.Pipeline.Generated;
 using Azure.Core;
